doc/cat-file: allow --use-mailmap for --batch options
The command git cat-file can now use the mailmap mechanism to replace
idents with their canonical versions for commit and tag objects. There
are several options like `-s`, `--batch`, `--batch-check` and
`--batch-command` that can be combined with `--use-mailmap`. But the
documentation for `-s`, `--batch`, `--batch-check` and `--batch-command`
doesn't say so. This patch fixes that documentation.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: John Cai <johncai86@gmail.com>
Helped-by: Taylor Blau <me@ttaylorr.com>
Signed-off-by: Siddharth Asthana <siddharthasthana31@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>