cocci & cache.h: apply variable section of "pending" index-compatibility
Mostly apply the part of "index-compatibility.pending.cocci" that
renames the global variables like "active_nr", which are a shorthand
to referencing (in that case) a struct member as "the_index.cache_nr".

In doing so move more of "index-compatibility.pending.cocci" to
"index-compatibility.cocci".

In the case of "active_cache_tree" we'd have a textual conflict with
"vd/skip-cache-tree-update" in "next"[1], and in the case of
"active_nr" with "ab/various-leak-fixes" in "seen". Let's exclude
those specific cases while moving the rule over from "pending".

1. dc5d40f5bc4 (read-tree: use 'skip_cache_tree_update' option,
   2022-11-10)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>