fixup! pack-bitmap.c: boundary-based walk
Reuse the existing rev_info struct when doing fill-in traversal on the
boundaries. This prevents us from issues like accidentally not filling
in trees/blobs by forgetting to set tree_objects appropriately.

Doing so moves us much closer to the realm of plausibility:

    ~/s/linux [nand] (master!) $ git.gh rev-list --objects master --not --tags --count --use-bitmap-index
    2728

but this is still higher than the desired 366.

For later: the count jumps from 2723 to 2728 when using this approach as
opposed to just setting the appropriate <type>_object flags. May be
worth investigating later.