fixup! pack-bitmap.c: boundary-based walk
Calling reset_revision_walk() doesn't clear the UNINTERESTING flags on
parsed objects, which cuts off the traversal in unexpected ways.

This resulted in an incomplete set of haves, which in turn gave us too
many results.

Make sure that we clear that flag before building up the haves bitmap.
This gives us the expected output, but we now spend 3 seconds doing
fill-in traversal from the boundaries, so something else funky is going
on here...